refactor(core): 🔨 split WindowManager into partial classes
